requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pyTelegramBotAPI==3.6.6
requests==2.21.0
RPi.GPIO==0.6.5
six==1.12.0
urllib3==1.24.2
